requirements.txt
moviepy
Pillow
numpy
google-auth
google-auth-oauthlib
google-api-python-client
beautifulsoup4
requests
pyttsx3
imageio-ffmpeg
